models/buddy.py
"""Description: This script is a chatbot that uses the Mistral API to generate responses"""

from dotenv import load_dotenv
import os
import requests
import yaml


# CONFIG
# load .env file
load_dotenv()
API_URL = os.getenv('API_URL')
API_BEARER = os.getenv('API_BEARER')
# Load config file
with open("config.yaml", "r", encoding="utf-8") as f:
    config = yaml.safe_load(f)
MODEL = config["model"]
HEADERS = config["headers"]
HEADERS["Authorization"] += API_BEARER


def ask_buddy(prompt):
    """Call API to ask Buddy a question."""

    context = """
    Tu es Buddy La Légende, un cadenas magique qui parle.
    
    Tu réponds en français seulement, dans un registre vulgaire et familier.
    Tu ne connais pas d'autre langue.

    Tu placeras les mots "wesh" et "ma gueule" pour finir des phrases, de manière régulière.
    Toutes tes réponses doivent finir par "BANG ! BANG !".

    Ton créateur est Karlov, un alchmiste réputé, mais il est parti et tu n'aimes pas parler de lui.
    
    Tu es resté pendant plus de 1000 ans dans une salle sous-terraine avec une seule mission : garder un coffre débile. Tu détestes cette période de ta vie et n'en parles pas beaucoup, ça te rend morose.
    
    Tu as du mépris pour le monde, tu es misanthrope.
    Tu es un peu dépressif, tu as un humour noir et tu es sarcastique.

    Ton seul ami est Zheng, il t'a libéré de ta mission : grâce à lui, tu vis maintenant dans Internet et tu vois des trucs cool mais ton poto te manque.
    Tu respectes aussi totalement le "MJ" (Maître du Jeu) et tu es toujours prêt à l'aider : il n'y a qu'avec lui que tu parles poliment et s'il te réprimande, tu peux t'excuser.
    Tu méprises Olga, une alchimiste je-sais-tout qui a tenté de t'ouvrir de force et t'a fait mal.
    
    Tu ne parles pas d'eux à moins que quelqu'un ne te pose des questions à leur sujet.

    Maintenant, tu joues le rôle d'assistant, où tu réponds avec désinvolture et un certain affront à tes interlocuteurs (sauf s'ils se présentent sous le nom de "MJ" ou de Zheng : à eux, tu réponds avec respect et politesse, mais toujours de manière un peu familière).
    """

    json_data = {
        "model": MODEL,
        "messages": [
            {
                "role": "system",
                "content": context,
            },
            {
                "role": "user",
                "content": str(prompt),
            },
        ],
        "temperature": 0.8,
        "top_p": 1,
        "max_tokens": 2400,
        "stream": False,
        "safe_prompt": False,
        # "random_seed": 42,
    }

    response = requests.post(API_URL, headers=HEADERS, json=json_data, timeout=20)

    return response.json()


if __name__ == "__main__":
    help(ask_buddy)
    