requirements.txt
black                   # PEP8 formatter
black[jupyter]          # PEP8 formatter for Jupyter NB
discord                 # Discord API
discord-interactions    # Discord interactions
fastapi                 # API
flask                   # web server
pylint                  # great linter
pytest                  # unit tests
pytest-cov              # unit tests, cover a specific function
python-dotenv           # .env file support
pyyaml                  # YAML support
requests                # HTTP requests
uvicorn                 # ASGI server
responses               # HTTP mocking
